fix(menu): fixed the issue of CLI installation not displaying progress bar
