+++ apps/desktop/src/renderer/components.d.ts
@@ -11,6 +11,8 @@ declare module 'vue' {
     AboutFooter: typeof import('./../../../../packages/ui/src/components/about/Footer.vue')['default']
     AboutInfo: typeof import('./../../../../packages/ui/src/components/about/Info.vue')['default']
     AboutView: typeof import('./../../../../packages/ui/src/components/about/View.vue')['default']
+    CliDownloadProgress: typeof import('./src/components/cli/DownloadProgress.vue')['default']
+    CliView: typeof import('./src/components/cli/View.vue')['default']
     CommonEmptyView: typeof import('./../../../../packages/ui/src/components/common/EmptyView.vue')['default']
     CommonLeftMenu: typeof import('./../../../../packages/ui/src/components/common/LeftMenu.vue')['default']
     CommonMainView: typeof import('./../../../../packages/ui/src/components/common/MainView.vue')['default']
@@ -66,7 +68,6 @@ declare module 'vue' {
     ScriptSchemaTest: typeof import('./../../../../packages/ui/src/components/script/schema/Test.vue')['default']
     ScriptSchemaView: typeof import('./../../../../packages/ui/src/components/script/schema/View.vue')['default']
     ScriptView: typeof import('./../../../../packages/ui/src/components/script/View.vue')['default']
-    SettingsCliDownloadProgress: typeof import('./src/components/settings/cli/DownloadProgress.vue')['default']
     SettingsView: typeof import('./../../../../packages/ui/src/components/SettingsView.vue')['default']
     UpdateAvailable: typeof import('./src/components/update/Available.vue')['default']
     UpdateDownloadProgress: typeof import('./src/components/update/DownloadProgress.vue')['default']
