+++ apps/desktop/src/renderer/src/components/cli/DownloadProgress.vue
