+++ apps/desktop/src/renderer/src/components/cli/useCLIInstall.ts
@@ -1,0 +1,21 @@
+const intallCliBtnLoading = ref(false)
+
+export function useCLIInstall() {
+  async function installCli() {
+    if (intallCliBtnLoading.value) return
+
+    try {
+      intallCliBtnLoading.value = true
+      await window.api.installCLI()
+    } catch (error) {
+      console.error(error)
+    } finally {
+      intallCliBtnLoading.value = false
+    }
+  }
+
+  return {
+    intallCliBtnLoading,
+    installCli,
+  }
+}
