+++ apps/desktop/src/renderer/src/pages/settings.vue
@@ -1,35 +1,10 @@
 <script setup lang="ts">
 import useSettingsService from '@database/services/SettingsService'
+import { useCLIInstall } from '../components/cli/useCLIInstall'
 
 const { settings } = useSettingsService()
 
-const intallCliBtnLoading = ref(false)
-
-async function installCli() {
-  try {
-    intallCliBtnLoading.value = true
-    await window.api.installCLI()
-  } catch (error) {
-    console.error(error)
-  } finally {
-    intallCliBtnLoading.value = false
-  }
-}
-
-const cliDownloadProgressVisible = ref(false)
-const cliDownloadProgressPercent = ref<number | null>(null)
-const cliDownloaded = ref(false)
-
-window.api.onInstallCLIStatus((_event, cliDownloadEvent) => {
-  const { status } = cliDownloadEvent
-  if (status === 'download-progress') {
-    cliDownloadProgressVisible.value = true
-    cliDownloadProgressPercent.value = cliDownloadEvent.data.percent
-  } else if (status === 'cli-downloaded') {
-    cliDownloaded.value = true
-    cliDownloadProgressVisible.value = false
-  }
-})
+const { intallCliBtnLoading, installCli } = useCLIInstall()
 </script>
 
 <template>
@@ -37,10 +12,5 @@ window.api.onInstallCLIStatus((_event, cliDownloadEvent) => {
     v-model="settings"
     :install-cli-btn-loading="intallCliBtnLoading"
     @install-cli="installCli"
-  />
-  <SettingsCliDownloadProgress
-    v-model="cliDownloadProgressVisible"
-    :percent="cliDownloadProgressPercent"
-    :update-downloaded="cliDownloaded"
   />
 </template>
